Add support for Installing Import Targets for TileDB. #596
This add support for installing cmake's TileDBConfig.cmake file for
find_package(TileDB) support. tiledb_shared is now privately linked to
dependencies to avoid setting the targets as dependencies when
tiledb_stared is used as imported target. tiledb_static now lists the
interface_link_libraries so importing its target requires the
dependent targets for linking.

The unit tests are switched from linking against the shared library to
the static library in order to pull in the linking requirements.

Documentation for usage was also updated with details and examples for
find_package() usage.