Disable 'S3ThreadPoolExecutor'
We're seeing occassional hangs in the unit tests.

From #s3:
When `VFS` destructs, it may destruct `thread_pool_` before `S3`. This leaves
the new S3ThreadPoolExecutor with a garbage 'vfs_thread_pool_' ptr. I would
normally expect some kind of segfault, but I wouldn't rule out a hang since the
behavior is technically undefined.

The bottom line is that I ran our unit tests hundreds of times (just using the AWS
`DefaultExecutor` instead our of `S3ThreadPoolExecutor`) without a hang. I think
for 1.5, we should just use the DefaultExecutor to reduce regression risk. I don't
see any of the locks or wait()s are hanging in the `S3ThreadPoolExecutor`, which
is why I suspect it is a lifetime issue.