Fix infinite loop in S3::ls() when the 'delimiter' is empty
From the documentation on Aws::String Aws::S3::Model::ListObjectsResult::GetNextMarker():
Note: This element is returned only if you have delimiter request parameter specified. If
response does not include the NextMaker and it is truncated, you can use the value of the
last Key in the response as the marker in the subsequent request to get the next set of
object keys.

So, when 1) the 'delimiter' attribute passed to S3::ls() is empty, and 2) the response
is truncated we get stuck in an infinite loop because we never set the correct marker
in the follow-up requests.