Refactor 'cancel_all_tasks' out of 'ThreadPool'.
The motiviation behind this change is that our ThreadPool::enqueue() routine
does not guarantee that the scheduled task will ever actually execute.

This can happen when ThreadPool::cancel_all_tasks() is invoked between the time
a task is enqueued and when one of the worker threads attempt to execute the task.

Currently, we check for cancellation by waiting-and-getting the returned std::future
from ThreadPool::enqueue(). For example:
std::future f = tp_.enqueue(...);
f.wait();
if (!f.get().ok()) {
  // It may have been cancelled
}

The limitation is that some asynchronous callers will never block on the returned
future, or may block using other methods. We saw this when attempting to integrate
our threadpool with the AWS SDK in S3ThreadPoolExecutor. Our tests were hanging
because the AWS SDK makes the assumption that all scheduled tasks would execute,
and we can not notify the SDK otherwise.

Consider this scenario:
std::semaphore sem(1);
tp_->enqueue([&sem]() { sem.relase(); });
// Some other thread invokes tp_->cancel_all_tasks(). This thread is hung.
sem.aquire();

To solve this, I've introduced a CancelableTasks class. This allows the caller to
schedule and track tasks scheduled among one or more threadpool instances. The
first use-case will be that we can implement the contract of
Aws::Utils::Threading::Executor::Submit() with the the 'VFS::thread_pool_', while
still allowing VFS::cancel_all_tasks() to avoid hanging the AWS SDK.

I've also expanded the ThreadPool::enqueue() contract to guarantee that the task
will eventually execute, if the returned std::future's ::valid() evaluates to true.
If it evalutes the false, it indicates that the task will never execute. I've
verified (or modified) that all existing uses of ThreadPool::enqueue() conform to
this contract.

Note that the S3ThreadPoolExecutor is still unused/disabled. I will enable it in
a follow-up patch.

I made other misc. changes that appeared to be vestiges of the C codebase (e.g.
using heap-allocated threadpools in storage manager, where a stack-allocated
class member makes more sense).

Tests run: make check