Modify incorrect assertion in S3::ls()
I assumed that part of the `ListObjects` AWS contract was that if the
response was truncated, it must contain at least one item. The documentation
is ambigious, but Isaiah was able to reproduce one of these empty-and-truncated
responses on AWS, GCS, and Minio when using a delimiter. This narrows the
scope of the assert to only make this assumption when a delimiter is
unspecified.