Use new cmake variables for libcurl linkage. #1080
Switch curl superbuild to build with cmake instead of autotools. This
create the cmake export targets.

When finding lib curl if the libcurl export exists, we use that, which
if libcurl is new enough includes the link dependencies, so libtiledb
can include them.

Note: I can't find a cmake equivalent of --enable-optimize