Query Submit V2, part 1 (#1380)
This is part 1 of 2 for modifying the core tiledb rest client to use the
new /v2/array/.../query/submit API. The differences between v2 and v1 are:

- For all query types, each returned serialized query (aka BufferList) is prefixed
with a little-endian ordered unsigned int that contains the total byte size of
the serialized query that follows it in the response buffer.
- When the 'read_all' request parameter is 'true', the server will continue
resubmitting incomplete queries until the request is complete. This guarantees
that the user will recieve a complete query request in the response.
- If the 'read_all' request required N number of query operations to
core tiledb on the server, the response buffer will be organized as follows:
<prefix><serialized query 1><size><serialized query 2>...<size><serialized query N>

This is part 1 of 2 because it does NOT:
- Process a buffered response
- Provide any sort of retry logic from the client.

In other words, this waits for the entire response to be received before it
starts processing (deserializing and copying results back to the user). The next
patch will make use of the curl library's `write_callback` to begin processing
the response as it receives it. I have yet to give any thought to how the
client will detect incomplete responses and appropriately restart a query
from the last receive serialized query object. I am not sure whether this
should be opaque or transparent to the user.

Note that this does make a copy for each serialized query to ensure 8-byte
alignment. This is not necessary if the buffer is already 8-byte aligned, but
this copy will be thrown away in favor of some sort of buffer pool for
handling the buffered response. I didn't bother trying to optimize this
path to prevent obfuscating the logic.