Query Submit V2, part 2
This is part 2 of the client-side handling of the /v2 query submit
requests.

This introduces a callback that the Curl object invokes as it
receives partial response data. The callback is defined by the
RestClient object.

As each complete, serialized query object is recieved, the client
will deserialze each object and copy the response into the 'copy_state'.
There are two primary benefits to this design:
- Reduce total operational latency. The client does not wait for the
  entire response to be recieved before processing it. It can also
  process the partial responses while the network is receiving the
  remaining response data.
- Reduced memory consumption. This patch discards the serialized
  query objects as soon as they are recieved and processed. Consider
  the current scenario where the client recieves an entire 2GB response
  and copies ~2GB of buffers back into the copy state. With this patch,
  the peak memory consumption in this path should be reduced ~50% because
  the buffered response data size should be negligible (~16KB).
- When 'resubmit_incomplete_' is set but the request returns an incomplete
  query, an error is set to alert the client that they should retry the
  returned query.

I also noted a few micro-optimizations in the comments of this patch.
These can be done in a part 3, but should not be required for releasing
1.7.

I also modified the query_deserialize() routine to respect the
internal offset of the buffer to deserialize.