Add runtime detection for CA Certs on Linux
This solves a major issue when we build libcurl as part of a super
build. Curl at configuration/compile time detects the systems CA bundle
path, and then hard codes this into the compiled lib. This causes issues
when libtiledb has a statically linked libcurl and is copied between
linux systems (i.e pip or maven). Often then the hard coded CA bundle
path is no longer valid and curl fails to handle SSL.

The solution is to at runtime try to detect the location of the CA cert
bundle for the host os.

Currently the CA cert issues only exists on linux, so we only do runtime
detection for linux.

[ch814]