Cache REST array schema
This saves about 500-1000ms on a REST query when using
the Python API. I've measured REST requests as quick as 2.2s
from my laptop with this patch.

The following line in my benchmark generates 3x requests to
get the array schema:
`with tiledb.open("tiledb://stavros/dense_X", ctx=tiledb.cloud.Ctx()) as A:`

This patch just caches the return value and re-uses it between the calls.
This is technical debt because the real fix will be to trace where the
3x calls are coming from and reduce it to 1. Some quick tracing revealed
1x of the calls are generated from `Array::open` and the other 2x are
called outside of the core at `tiledb_array_schema_load_with_key`.