WIP: Hide Tile::data() and Tile::cur_data()
This removes the Tile::data() and Tile::cur_data()
interfaces. Where possible, we now use Tile::read() --
which is not thread-safe. Within multi-reader scenarios,
e.g. the callback routines within `parallel_for` that
currently invoke Tile::data(), we must use a thread-safe
variant of Tile::read() which is now defined as:
Tile::read(void* buffer, uint64_t nbytes, uint64_t offset)

This is a WIP because there are failures in the unit test.
I'd prefer to have the general approach reviewed before I
started debugging the unit test failures.

Also not that the ResultCoords object currently takes a raw
pointer to an address of a Tile's internal buffer. Since this
is no longer exposed, we're now making a copy for
ResultCoords::coords_. To minimize code churn, I've converted
this raw pointer into a shared_ptr<T>, which may contain
multiple T objects. This is a little bit of an anti-pattern
because we should be using some type of container, e.g.
shared_ptr<vector<T>>. We may not need the shared pointer at all,
but I didn't want to risk regression.