WIP: Add support for lazy loading remote array's metadata
This makes changes so remote arrays are able to lazy load their
metadata.

A few const had to be removed as now several previously readonly
functions might cause various pieces of the metadata to be loaded.

The metadata_index used for index based lookups is now built on the fly
if the index is empty when fetching a metadata value by index id.