Do no use stdout as default file in dump() interfaces (#1482)
It triggers a (harmless, but annoying as consistent) warning from the R
'linter' (ie R CMD check) as output for code called from R should go through
R's buffered stream (which is reasonable and commonly done).

So we default to nullptr and if nullptr is seen, switch to stdout in the
function body rather than interface. This way the R checks are 'clean' which
is preferable, and at no real cost to the library layer (apart from four if()
statements in code off any hot paths).