Refactor error handling on S3::is_object and S3::is_bucket
Currently, S3::is_object and S3::is_bucket can return false negatives. They
can return 'false' in the scenario where the object/bucket exists but the http
request failed for other reasons. This patch does not fix that. This patch only
refactors the interfaces so that they return both a `Status` and `bool` instead
of only returning a `bool`.

The intent of this patch is to handle the refactoring so that the follow-up
patch will only touch the implementations of `is_object` and `is_bucket`.