Add cmd line arg '--vfs' to 'tiledb_unit'
Currently, the unit tests determine which VFS filesystem from the build context
of the core library. For example: If the core library was built with "--enable-hdfs",
unit tests will use the HDFS VFS implementation. This command line argument
"--vfs" allows the user to override the VFS filesystem.

`./tiledb_unit --vfs native` will use either POSIX/Win VFS FS, even if the
core was built with S3 and/or HDFS support.

Note that the above will still run the S3 and/or HDFS specific unit tests. If
the user built with HDFS but does not wish to run ANY tests with HDFS, they
will need to use `./tiledb_unit ~[hdfs] --vfs native`. This is required because
the flag does not override the compile-time building of the FS-specific unit
tests. This seems OK since this shouldn't be commonly used.

Note that the command line argument now shows up with the `-h` flag.

```
./tiledb/test/tiledb_unit -h

Catch v2.2.1
usage:
  tiledb_unit [<test name|pattern|tags> ... ] options

where options are:
  ...
  --vfs <native|s3|hdfs>                    Override the VFS filesystem to
                                            use for generic tests

For more detailed usage please see the project docs
```