FileSystem::Azure chunked upload (#1498)
This provides a chunked upload implementation of Azure::write(). The chunk
uploads are not performed in parallel -- that will be done in a future patch.

The immediate benefit to this patch is that a user can write large blobs
without buffering it entirely in-memory.

The interface allows for a user to invoke Azure::write() as many times as they
need to without commiting anything to Azure. When Azure::flush_blob() is invoked,
either all written buffers are commited as a single blob or none of them are.

The high-level algorithim identical to FileSystem::S3, which is:
- Azure::write() will cache writes in-memory until exceeding a predefined maximum
  size.
- When the size of the in-memory write cache size exceedes the maximum size,
  Azure::write() will block while flushing the cached writes to Azure as
  uncommited blocks.
- When Azure::flush_blob() is called, the in-memory write cache will flush the
  remaining bytes in the write cache. Next, it will send a request to Azure to
  commit all of the uncommited, uploaded blocks as a single blob.

Since this is going into a dev branch, don't worry about the in-source TODOs
for now.