CI: Build with Azure FS support for Linux and OS X (#1508)
This introduces building with '--enable-azure' to the CI for Linux and OS X.

Note that this does not put any of the Azure FS code under test. This patch
disables the [azure] tagged unit tests and does NOT modify our unit test
helpers to use the Azure VFS implementation for generic unit tests.

This would be a relatively straight forward patch if not for the Azure SDK's
dynamic linking dependency on libuuid. Our core library can build independently
without linking with libuuid, but executable builds fail that link with
libtiledb without linking with libuuid. I've modified some of our example
builds to link with libuuid. I've also modified the CI to install uuid-dev
through apt on Linux and to install ossp-uuid through homebrew on OS X.

The libuuid dependency is up for discussion. We don't actually use it in the
code so we could 1) hack around it implementing our own no-op implements of
the two undefined routines in libtiledb, 2) fork our own Azure SDK and remove
the dependency, or 3) fetch the source, build it, and perform a static link
into libtiledb.