Patch out uuid-dev dependency in the Azure SDK (#1511)
The Azure SDK depends on the uuid-dev library, but we will never path through
the uuid-dev routines in our usage of the Azure SDK. This patch provides a
patch-file that we apply during the CMake process. This introduces a maintence
problem if the patch ever conflicts with an updated version of the upstream
repo. I'm OK with this because the patch file is tiny and we will be rarely
updating the Azure SDK.