Context-free "filtered" keyword semantics
Currently, both the Tile::filtered_ and the query context (e.g. in the read or
write path) are required to determine if the Tile::buffer_ contains the on-disk
format. For example: when 'filtered_' is true in the write path, 'buffer_'
contains the on-disk format. When 'filtered_' is true in the read path,
'buffer_' does not contain the on-disk format.

This patch changes the semantics of the "filtered" to mean "filtered, on-disk
formatted" and "unfiltered" to mean "reverse-filtered, in-memory formatted".

This may be a useful cleanup, but the motiviation for this patch is to prepare
for the vectorized tile buffer patch:

'void* Tile::buffer_' will be removed in favor of 'ChunkedBuffer*
Tile::chunked_buffer_' and 'void* Tile::filtered_buffer_'. The 'chunked_buffer_'
contains the "reverse-filtered, in-memory formatted" data. The 'filtered_buffer_'
contains the "filtered, on-disk formatted" data.

With the new semantics of the "filtered" keyword, 'bool Tile::filtered_' becomes
unneccessary because "bool Tile::filtered()" can be implemented by determing
which of the two buffers ('chunked_buffer_' or 'filtered_buffer_') are non-empty.

In other words: the state of Tile will become more complex to support two
separate buffers. This patch is a pre-requisite to reducing some of that state
by allowing us to eliminate "bool Tile::filtered_".