Add checksum filters for MD5 and SHA256 [ch1600]
These filters will compute checksums on the bytes and metadata being
written and store the results in the chunk metadata. On reads the
checksums of all metadata and data parts are validated to ensure the
data read is as expected.

The checksum filters are only helpful to validate reads are getting the
correct bytes from storage. The filters do not help to validate a write
was indeed succesful. That will be left for an additional fragment level
checksum.