Store Filtered Tiles in the Tile Cache
In preparation for selective unfiltering, we can no longer store unfiltered
tiles in the tile cache because the unfiltered tiles may be incomplete. This
patch stores in filtered tiles in the tile cache instead of unfiltered cache.

In it's current state, this trades lower memory consumption in the tile cache
for higher CPU consumption when unfiltering cached tiles.