VFS path behavior unification (#1541)
This is a first step toward ensuring that the way we handle and report
paths are consistent among our VFS backends.

This patch introduces a unit test that ensures ls() and touch() are consistent
among the VFS backends. In the future, we can continue expanding this unit test
to test and correct other discrepancies. We'll use the posix behavior as out
target behavior.

A few changes:
- Ensures paths returned from ls() are without a trailing slash.
- Ensures the URI constructor does not drop a trailing slash in the scenario
  where a relative path is converted to an absolute path.
- Ensures touch() fails if the given URI has a trailing slash.