Azure FS: Conditional retry on HTTP code
This modifies the retry path so that we only perform a retry on a server-side
error.

Consider the scenario where we want to check if an object exists. If the
object does not exist, we will receive a 404 and retry 2 more times with a
1 second delay between them.

With this patch, we will return after receiving the first 404.