GCS VFS: Defer zero-copy buffers
The GCS SDK uses std::string objects to represent buffers within the IO
calls. The interfaces force the std::allocator<char> for the std::string so
it is not possible to override the allocator to eliminate excessive copies
in our VFS call stack.

I spent some time trying to patch the GCS SDK to allow for a custom allocator
(or just a `std::pair<char *, size_t>`) but it is not trivial.