Thread-safe Catch2 Macros: (#1590)
The concurrent write units were occassionally crashing with a segfault. The
trace in the core indicated that the Catch2 REQUIRE macro was not thread safe.

I confirmed this with the public documentation here:
https://github.com/catchorg/Catch2/blob/master/docs/limitations.md#thread-safe-assertions

For this patch, I've introduced thread-safe variants of the REQUIRE and CHECK
macros: REQUIRE_SAFE and CHECK_SAFE. These macros use a global mutex to protect
the Catch2 macro that they wrap.