Benchmarks for Selective Unfiltering
This introduces two benchmarks that I've used while verifying selective
unfiltering.

1. bench_large_io.cc
This benchmark reads/writes to dense/sparse arrays with a large tile extent
size. Both arrays have fixed and var-sized data.

With selective unfiltering, this runs in ~83.5 seconds. Without selective
unfiltering, this runs in ~81 seconds. I interpret this as a ~3% slowdown due
to overhead in 1) vectorizing the value tiles and 2) performing the intersection
logic to determine if a tile should be selectively unfiltered.

2. bench_selective_unfiltering.cc
This benchmark creates a dense array with fixed and var-sized data. The benchmark
section only records runtime and memory statistics for 20 serial reads, where
each reach is a multi-range query designed to intersect each tile in the array
but only one chunk. Each tile has roughly 40 chunks. Tile cache disabled.

This benchmark was sufficient to determine that both peak and average
memory is reduced with selective unfiltered. Here are the results:

// Selective unfiltering
“runtime_ms”: “6784"
“peak_mem_mb”: “206"
“avg_mem_mb”: “177"

// Without selective unfiltering (aka dev branch)
“runtime_ms”: “14901"
“peak_mem_mb”: “367"
“avg_mem_mb”: “265"

The above are single samples and not an average. The runtimes seem to vary
+/-20%, but the memory usage remains nearly identical between runs.