Add support for environmental variables for config
Environmental variables are now supported for setting config options.
The prefix of `TILEDB_` is used by default but is configurable with
`config.env_var_prefix`. Config parameters are converted from `a.b.c`
syntax to `A_B_C`. So for a parameter of `vfs.num_threads` the
environmental variable of `TILEDB_VFS_NUM_THREADS` is looked for.

The order we look for values are:
  1. user set config parameters
  2. env variables
  3. default config value