Fix overriding buffer w/split buffs and coords
In compatibility with coords and split buffers we need to properly shift
the buffer we are copying into. We also need to account for the position
argument that is based into the reader.

This was missed in the backwards compatibility tests because they only
fetch a single value from the array.

Unfortuantly not all compatibility test have more than one value in the
array, so we are unable to adjust the unit tests at this time. We need
to modify the compatiblity arrays to all have 2+ cells worth of data.
This will be done in a follow up PR at a later time.