Allow empty values for variable length attributes
The ascending order check for offsets is relaxed to allow two or more
succeeding values to have the same offset. This allows for a zero length
value to on one more cells of a variable length attribute.

Currently writting a single empty value is not supported. Empty cells
must be part of a write that includes at least 1 cell with values.

Consider the following examples:

Write to 4 cells (1-4) where the 4th cell of a1 is empty

d1 = {1, 2, 3, 4}
a1_data = {1, 2, 3}
a1_offsets = {0, 1, 2, 3}

write to 4 cells (1-4) where the 2nd cell of a1 is empty

d1 = {1, 2, 3, 4}
a1_data = {1, 2, 3}
a1_offsets = {0, 1, 1, 2}