Robust retries for S3 request-limit retries
This patch provides a retry handler that is identical to the existing, default
handler with an exception for CoreErrors::SLOW_DOWN. For this error, we will
unconditionally retry every 1.25-1.75 seconds.

The motivation for this patch is to allow the TileDB client to remain functional
even when performance may be bottlenecked on this error. The server returns
this error when we exceed a fixed number of requests per second. The client
will eventually make progress.