Improve GCS build
This adds a new patch to the GCS build which disables tests. The patch
is being upstreamed so hopefully we'll be able to drop it eventually.
This also adjust the installation and configuration of GCS to install
directly into the TILEDB_EP_INSTALL_PREFIX folder. This resolves errors
with the previous copying of `lib` vs `lib64`.