Always check ranges in read for multiple fragments
If there are multiple ranges and these ranges are satisfied by a single
tile inside a fragment we might incorrectly leave later ranges as being
from a single fragment. With overlap we were only counting a
fragment+tile for the first range we saw it for. Thus we might miss that
a following range was included in more than one fragment. This case
happened when a use consolidated the array but did not vacuum. This
results in two tiles satifying a range.

If a range is marked as being only from a single fragment the
deduplication process is not run, and multiple records could be
returned.