Reduce branching in coordinate sort path
This optimizes the coordinate sort path in scenarios where _all_ coordinates
are unzipped.

Currently, the ResultTile::coord() routine branches to determine whether it
is fetching a zipped or unzipped coordinate. This patch overloads the routine
with a `template <bool>` signature, where the template parameter determines
whether the coordinate is unzipped or potentially zipped. When the parameter
is false, it skips the branch and assumes the coordinate is unzipped. When
the parameter is true, it checks (so: unchanged behavior).

When one or more of the fragments are unzipped, the behavior is unchanged and
there is not a performance improvement.