ResultTile::coord() performance
This removes the branch within the ResultTile::coord() implementation to
test if it contains zipped or unzipped coordinates.

The function contract is slightly modified because now the caller must ensure
that an underlying coordinate exists at there requested position and dimension.
Currently, the `ResultTile::coord()` may return null if nothing is found. This
does not appear to be an expected output anywhere in the unit tests or code.

In a certain read benchmark that paths through this routine multiple times,
this patch reduces the avg execution time from 18.15s to 17.00s.