Check ENV for config values with config_iter
This fixes a problem where in python default values are printed instead
of the user set ENV config parameters. The iterator still only iterates
over set or defaulted config parameters. We don't iterate over all env
variables at this point, that is left to a future change if a user
requests it.