Disable sub-partitioner by default
This adds a configuration parameter "sm.sub_partitioner_memory_budget" that
defaults to 0. When 0, we will bypass the sub-partitioner. I expect this to be
an advanced tuning option unless we can, in the future, programmatically
determine if a workload will benefit from the sub-partitioner (e.g. the recent
dataframes example).