Fix compute_mbr for var-sized dimensions
There is a mistake in the logic of `compute_mbr` and `compute_mbr_var` that
assumes that the var-sized cells in the value tile do not span more than one
chunked buffer. If a var-sized cell contains 2 integer values, it is only true
that the bytes for an individual integer value do not span a chunked buffer.
The first integer value may be at the end of one chunked buffer and the second
integer may be on the next chunked buffer.

This could be easily fixed, but I also noticed that the write path (where these
functions are called) guarantees that tile chunked buffers will always be
contiguous. In this scenario, we can just revert both `compute_mbr` and
`compute_mbr_var` to the logic before the chunked buffer patch (commit hash
4b4b502a0d27e8dfd51df674527df31a271caac8).