Sort cell slab ranges for ND arrays
The selective decompression intersection algorithm requires cell slab ranges
to be sorted in ascending order. This is true for 1D arrays, but not for ND
arrays. In this scenario, we must sort them.

The ranges are sorted in vectors that have already been partitioned per-tile.
This should keep the sort runtimes relatively quick. In the future, we can
benchmark this on large arrays+queries and measure its timing. With this sort,
we now may coalesce cell ranges as a future optimization.

If the N*LOG(N) sorting is too slow, the alternative approach is to leave them
unsorted and perform O(N*M) range-chunk intersection comparisons, where M is
the number of chunks. If the number of chunks is less than LOG(N), this may
be faster.

This solves the following error message:
[TileDB::ChunkedBuffer] Error: Chunk read error; chunk unallocated error