Coalesce Subarray Ranges
This patch will coalesce contiguous ranges with discrete data types at the
start of computing tile overlap. This ensures that ranges are coalesced
before they are used in both the read path and result size estimation path.
Additionally, tile overlap is only available for recompute if the ranges change.
This ensures that we do not needlessly rerun the computation to coalesce the
same ranges.

I am currently testing a multi-range index read query. A query that reads an
entire 40MB 2D dense array with a single range in 4 seconds. A query that
reads the same array with 10mil ranges runs in 40 seconds. With this patch,
the 40 seconds are reduced to 4.5 seconds. So computing and coalescing 10 mil
ranges takes about a half second, but saves 36 seconds in my scenario.