Recursive ThreadPool
This patch allows recusive usage of a single ThreadPool instance without deadlock.
While this is a pre-requisite for using a ThreadPool to replace the TBB scheduler,
it does provide a minor but immediate optimization for achieving the same level
of concurrency with one less OS thread.

Consider the following deadlock scenario with our current implementation of
ThreadPool:

```
ThreadPool tp;
tp.init(1 /* num_threads */);

auto outer_task = tp.enqueue([&]() {
  auto inner_task = tp.enqueue([]() {
    foo();
  });
  tp.wait_all_tasks(inner_task);
});

tp.wait_all_tasks(outer_task);
```

In this scenario, we hang on `tp.wait_all_tasks(inner_task);`. The `outer_task`
is running on the single worker thread within `tp`. The `out_task` makes recusive
usage of `tp` to enqueue `inner_task`. The single worker thread will never
make progress waiting on `inner_task` because it can not be scheduled.

This patch changes the behavior of the `ThreadPool:wait_all*()` routines. As
these routines iterate through their tasks to wait on, they will work on
pending tasks if the task they are waiting on has not completed. This ensures
they will never deadlock.

In the above scenario, the call to `tp.wait_all_tasks(inner_task);` would make
progress by executing `inner_task` before waiting on it.

Note that the ThreadPool::init() now accepts a `concurrency_level` instead of
`num_threads`. The motiviation for this is that I expect `tp.init(1)` to be
serial. In this scenario, there are no worker threads. All enqueued work
tasks are executed immediately.