Rename ThreadPool::enqueue to ThreadPool::execute (#1758)
This is a non-functional change that just renames ThreadPool::enqueue to
ThreadPool::execute. The motiviation for this patch is:
- We don't use a traditional FIFO queue for scheduling tasks anymore (it is
LIFO).
- The 'enqueue' name gives the impression that the tasks are always queued for
future execution. With a concurrency_level == 1, tasks may execute immediately.
This is an important distinction for reasoning about how locks behave. Consider
the following:

```
std::mutex my_mutex;
my_mutex.lock();
my_tp.enqueue([&]() {
  my_mutex.lock();
  // critical section
  my_mutex.unlock();
});
my_mutex.unlock();
```

In this snippet, we will deadlock if `enqueue` executes on the calling thread.
While this is contrived example and there are easy workarounds (for example,
using a std::recursive_mutex or releasing the lock before calling ::enqueue),
the `enqueue` name is misleading.

Recall that `wait_all*` has similar behavior (may execute tasks on the calling
thread) which is a necessary requirement for recursive use of a single ThreadPool
instance.

Co-authored-by: Joe Maley <joe@tiledb.com>