Coalesce Subarray Ranges (#1755)
For ranges of discrete data types, this patch will attempt to coalesce ranges
as they are added to a subarray. Ranges will be coalesced with the last existing
range if they form a contiguous raange.

Additionally, this introduces `bool Subarray::coalesce_ranges_` which controls
whether or not the instance will attempt to coalesce ranges as they are added.
This is on by default, but exists as a way for unit tests to disable them. The
motiviation it prevent the need for immediately refactoring all of the hard-coded
ranges in the subpartitioner unit tests. Subarrays split or cropped from an
existing subarray will be use the same coalescing behavior as the parent.

Co-authored-by: Joe Maley <joe@tiledb.com>