Remove use of TILEDB_COORDS in cpp-api
The TILEDB_COORDS macro invokes `tiledb_coords()` which is tagged for
compile-time deprecation. To avoid compile-time warnings when importing the
cpp-api, we must not use the TILEDB_COORDS macro. This patch hard-codes the
"__coords" name. This is reasonably safe because the named-coordinate attribute
exists for backwards compatability and will not be changed in the future.