Recursive ThreadPools (#1774)
Currently, ThreadPool instances are only recursive with their own isntance. This
patch allows multiple ThreadPool instances to recurse among themselves without
the potential for bottlenecking (either in performance reduction or a deadlock)
on available threads.

For instance, ThreadPool instances A and B may be called in any order:
A->execute([&]() {
  B->execute([&]() {
    A->execute([&]() {
      foo();
    });
  });
});

The motiviation for this patch is to allow us to use our `compute_tp` and
`io_tp` without worrying about their order in a call stack. This patch reduces
the runtime of our `bench_large_io` from 280587ms to 269621.

See the new unit tests in `unit-threadpool.cc` for more examples.

---

Design and Implementation:

Currently, executing a task on a ThreadPool returns an `std::future`. This patch
modifies the return value to a new ThreadPool::Task object. This object is
a relatively opaque handle that forces the caller to use the ThreadPool::wait_*
routines.

The motivation for this patch is to allow the ThreadPool to wake up a thread
blocked on ThreadPool::Task::wait() in scenarios where the task has not
completed. It is not possible to wake up a thread blocked on an std::future::wait().

Within the ThreadPool::execute() task, it now notifies a blocked ThreadPool::task
in the ThreadPool::wait*() path to wake up and check the call stack.

Co-authored-by: Joe Maley <joe@tiledb.com>