VFS Read-Ahead Cache
This introduces a read-ahead cache within VFS::read(). This is an LRU cache
that maintains a single cached buffer for an arbitrary number unique URIs, not
to exceed 10MiB (by default). Each cached buffer has a max size of 100KiB (by
default). These parameters can be tweaked with the following config items:

`vfs.read_ahead_size` (100KiB default)
`vfs.read_ahead_cache_size` (10Mib default)

The motiviation for this patch is to optimize IO patterns of small, relatively
sequential reads against cloud storage backends. Only the S3, Azure, and GCS
backends utilize this read cache. The POSIX/Windows/HDFS backends are
unaffected by this patch.

Both performing and caching the read-ahead incur a performance penalty:
1. We must read more than the requested bytes.
2. We must make a copy of the read buffer (one to store in the cache, one to
return to the user).

We will only perform a read-ahead if the requested read is smaller than the
default 100KB cached buffer size. IO patterns of large reads will be unaffected.
The assumption is that fragment data is large and that reading fragment
data will not incur a performance penalty. Additionally, reads to tile data
are bypassed because tiles have their own separate tile cache.

On the recent S3 workload we've been discussing, this read cache has a 78%
hit rate, where every cache hit is in the fragment metadata. I've observed a
best-case runtime of 6.5s with this patch, and a 27s runtime without this
patch.